+++ src/main/java/pl/asie/computronics/integration/railcraft/IntegrationRailcraft.java
@@ -2,7 +2,7 @@ package pl.asie.computronics.integration.railcraft;
 
 import cpw.mods.fml.common.registry.GameRegistry;
 import net.minecraft.block.Block;
-import pl.asie.computronics.Computronics;
+import net.minecraftforge.common.config.Configuration;
 import pl.asie.computronics.block.BlockDigitalDetector;
 import pl.asie.computronics.block.BlockDigitalReceiverBox;
 import pl.asie.computronics.block.BlockLocomotiveRelay;
@@ -20,9 +20,13 @@ public class IntegrationRailcraft {
 	public ItemRelaySensor relaySensor;
 	public Block digitalBox;
 
-	public IntegrationRailcraft(Computronics computronics) {
+	private static boolean isEnabled(Configuration config, String name, boolean def) {
+		return config.get("enable.railcraft", name, def).getBoolean(def);
+	}
 
-		if(computronics.isEnabled("railcraftLocomotiveRelay", true)) {
+	public IntegrationRailcraft(Configuration config) {
+
+		if(isEnabled(config, "locomotiveRelay", true)) {
 			locomotiveRelay = new BlockLocomotiveRelay();
 			GameRegistry.registerBlock(locomotiveRelay, "computronics.locomotiveRelay");
 			GameRegistry.registerTileEntity(TileLocomotiveRelay.class, "computronics.locomotiveRelay");
@@ -30,12 +34,12 @@ public class IntegrationRailcraft {
 			relaySensor = new ItemRelaySensor();
 			GameRegistry.registerItem(relaySensor, "computronics.relaySensor");
 		}
-		if(computronics.isEnabled("railcraftDigitalReceiverBox", true)) {
+		if(isEnabled(config, "digitalReceiverBox", true)) {
 			this.digitalBox = new BlockDigitalReceiverBox();
 			GameRegistry.registerBlock(digitalBox, "computronics.digitalBox");
 			GameRegistry.registerTileEntity(TileDigitalReceiverBox.class, "computronics.digitalBox");
 		}
-		if(computronics.isEnabled("railcraftDigitalDetector", true)) {
+		if(isEnabled(config, "digitalDetector", true)) {
 			detector = new BlockDigitalDetector();
 			GameRegistry.registerBlock(detector, "computronics.detector");
 			GameRegistry.registerTileEntity(TileDigitalDetector.class, "computronics.detector");