+++ changelog.md
@@ -3,8 +3,8 @@
   * FIXED: Updated Railcraft integration (requires Railcraft 9.5 now)
   * CHANGED: Some changes to the config file
     - It is recommended to regenerate the config file
-  * CHANGED: Got rid of the EnderIO CapacitorBank's getEnergyStoredForNetwork and getMaxEnergyStoredForNetwork methods
-    - Their behaviour was the same is getEnergyStored/getMaxEnergyStored, so use those methods instead
+  * CHANGED: Got rid of the EnderIO CapacitorBank's getEnergyStoredForNetwork and getMaxEnergyStoredForNetwork functions
+    - Their behaviour was the same is getEnergyStored/getMaxEnergyStored, so use those functions instead
 
 ==== 1.4.0 ====